+++ src/app/invitations/invitation-form/invitation-form.component.html
@@ -48,8 +48,9 @@
             placeholder="Select Teams"
             [styleClass]="'w-full'"
             emptyMessage="No teams found in this organization"
-            i18n-placeholder
             display="chip"
+            i18n-placeholder
+            i18n-emptyMessage
           />
         </div>
         <div class="field col-11 md:col-1 md:max-w-min min-w-min">
